+++ src/main/java/chess/game/Board.java
@@ -1,10 +1,14 @@
-package chess;
+package chess.game;
 
-import static chess.Position.of;
+import static chess.position.Position.of;
 import static chess.piece.Color.BLACK;
 import static chess.piece.Color.WHITE;
 
 import chess.piece.*;
+import chess.position.Column;
+import chess.piece.Direction;
+import chess.position.Position;
+import chess.position.Row;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
