+++ src/main/java/chess/position/Row.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.position;
 
 import java.util.Arrays;
