+++ src/main/java/chess/status/Running.java
@@ -3,8 +3,8 @@ package chess.status;
 import static chess.piece.Color.BLACK;
 import static chess.piece.Color.WHITE;
 
-import chess.Board;
-import chess.MoveCommand;
+import chess.game.Board;
+import chess.game.MoveCommand;
 import chess.piece.Color;
 import chess.view.Command;
 import java.util.Map;
