+++ src/test/java/chess/game/BoardTest.java
@@ -1,11 +1,13 @@
-package chess;
+package chess.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.game.Board;
+import chess.game.MoveCommand;
 import chess.piece.Color;
-import chess.piece.Queen;
 import chess.piece.Rook;
+import chess.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
