+++ src/test/java/chess/piece/BishopTest.java
@@ -3,7 +3,7 @@ package chess.piece;
 import static chess.piece.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.MoveCommand;
+import chess.game.MoveCommand;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
