+++ src/test/java/chess/piece/PawnTest.java
@@ -4,7 +4,7 @@ import static chess.piece.Color.BLACK;
 import static chess.piece.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.MoveCommand;
+import chess.game.MoveCommand;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
