+++ src/test/java/chess/status/FinishedTest.java
@@ -3,7 +3,7 @@ package chess.status;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.MoveCommand;
+import chess.game.MoveCommand;
 import chess.view.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
